Implementation of class StorageZone
Changed name from Zone to StorageZone. StorageZone has a list of aisles. An aisles storage specification is now decided by the storage zone that it is in. When adding a packing to an aisle it now checks if the storage specification of the package is the same as the storage specification of the storage zone that the aisle is in.